+++ src/app/tasks/page.tsx
@@ -1,12 +1,16 @@
+"use client";
+
 async function TasksPage() {
-    const response = await fetch('http://localhost:3000/api/tasks', {
-        cache: "no-store",
-    });
-    const tasks = await response.json();
+    // const response = await fetch("http://localhost:3000/api/tasks", {
+    //   cache: "no-store",
+    // });
+    // const tasks = await response.json();
+  
+    // console.log("tasks:", tasks);
 
-    console.log('tasks', tasks);
-
-    return <div>TasksPage</div>
-}
-
-export default TasksPage;
+    console.log('hey from a client component ');
+  
+    return <div>TasksPage</div>;
+  }
+  export default TasksPage;
+
